Quality improvements (#45)
- Enabled Windows build support
- Fixed a bug where GetSetAttribute<T> does not correctly cast `this` to `ASTNodeAttribute`
- Enabled `__PRETTY_FUNCTION__` support in MSVC
- Added [[nodiscard]] in `ASTNodeVisitor`
- Added casts to fix warnings of sizet->int, also fixed bug of incorrect width value (-1)
- Cleaned up `#include`
- Added comments in test HRML files

+++ compiler/parser/src/ASTBuilder.cpp
@@ -268,7 +268,7 @@ void ASTBuilder::visit(SubprocDefinitionPTNodePtr node)
         auto param_name = param->get_value();
 
         param_node = std::make_shared<VariableDeclarationASTNode>(
-            param->lineno(), param->colno(), param->lineno(), param->colno() + param->get_token()->width(),
+            param->lineno(), param->colno(), param->lineno(), param->colno() + static_cast<int>(param->get_token()->width()),
             param_name,
             nullptr);
     }
@@ -288,7 +288,7 @@ void ASTBuilder::visit(FunctionDefinitionPTNodePtr node)
         auto param_name = param->get_value();
 
         param_node = std::make_shared<VariableDeclarationASTNode>(
-            param->lineno(), param->colno(), param->lineno(), param->colno() + param->get_token()->width(),
+            param->lineno(), param->colno(), param->lineno(), param->colno() + static_cast<int>(param->get_token()->width()),
             param_name,
             nullptr);
     }
